requirements.txt
Flask==3.0.2
pyserial==3.5
Werkzeug==3.0.1 